20230201-solved-P150365: solving this problem using case simulation
Using brute force causes a timeout.
Not as much as 4^2500 because of the range, but a very large number of cases occurs.

We want the results in alphabetical order, so we search in "dlru" order.
If the moving distance is very large during search, it can be consumed by repeating "rl".

It is impossible if the initially given moving distance is shorter than the distance between the starting point and the destination, or if the remaining distance is not an even number.